Thwart virus scanners in email flow
