+++ src/config.rs
@@ -106,8 +106,8 @@ pub struct Templates {
     pub error: Template,
     /// A dummy form used to redirect back to the RP with a POST request.
     pub forward: Template,
-    /// A dummy form used to capture fragment parameters.
-    pub fragment_callback: Template,
+    /// A dummy form used to capture query and fragment parameters.
+    pub rewrite_to_post: Template,
 }
 
 
@@ -127,7 +127,7 @@ impl Default for Templates {
             login_hint: Self::compile_template("tmpl/login_hint.mustache"),
             error: Self::compile_template("tmpl/error.mustache"),
             forward: Self::compile_template("tmpl/forward.mustache"),
-            fragment_callback: Self::compile_template("tmpl/fragment_callback.mustache"),
+            rewrite_to_post: Self::compile_template("tmpl/rewrite_to_post.mustache"),
         }
     }
 }
