+++ src/handlers/mod.rs
@@ -1,3 +1,4 @@
 pub mod auth;
 pub mod normalize;
 pub mod pages;
+pub mod rewrite_to_post;
