--- tmpl/fragment_callback.mustache
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Portier &ndash; Forwarding...</title>
-    <script src="/static/fragment_callback.js" defer></script>
-  </head>
-  <body>
-    <form id="form" action="/callback" method="post"></form>
-  </body>
-</html>
