+++ tmpl/rewrite_to_post.mustache
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Portier &ndash; Forwarding...</title>
+    <script src="/static/rewrite_to_post.js" defer></script>
+  </head>
+  <body></body>
+</html>
